requirements.txt
confluent-kafka==2.6.0
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
fonttools==4.54.1
joblib==1.4.2
kafka==1.3.5
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
